+++ src/pages/toeic/Listening.jsx
@@ -4,7 +4,7 @@ import SEOHead from '../../components/SEOHead';
 import { useLanguage } from '../../contexts/LanguageContext';
 import useAOS from '../../hooks/useAOS';
 import SubNav from '../../components/SubNav';
-import SectionNav from '../../components/SectionNav';
+import PageLayout from '../../components/PageLayout';
 import { listeningQuestions } from '../../data/toeicData';
 
 const partOverviews = [
@@ -122,9 +122,8 @@ function Listening() {
       </section>
 
       <SubNav category="toeic" />
-      <SectionNav sections={sections} />
-
-      <div className="content-page">
+      <PageLayout sections={sections}>
+        <div className="content-page">
 
         {/* Part Overviews */}
         <div id="parts">
@@ -275,6 +274,7 @@ function Listening() {
           </div>
         </section>
       </div>
+      </PageLayout>
     </>
   );
 }