+++ src/pages/writing/BasicSentence.jsx
@@ -4,7 +4,7 @@ import SEOHead from '../../components/SEOHead';
 import { useLanguage } from '../../contexts/LanguageContext';
 import useAOS from '../../hooks/useAOS';
 import SubNav from '../../components/SubNav';
-import SectionNav from '../../components/SectionNav';
+import PageLayout from '../../components/PageLayout';
 
 const sentenceStructures = [
   {
@@ -193,9 +193,8 @@ function BasicSentence() {
       </section>
 
       <SubNav category="writing" />
-      <SectionNav sections={sections} />
-
-      <div className="content-page">
+      <PageLayout sections={sections}>
+        <div className="content-page">
 
         {/* Sentence Structures */}
         {sentenceStructures.map((structure, index) => (
@@ -294,6 +293,7 @@ function BasicSentence() {
           </div>
         </section>
       </div>
+      </PageLayout>
     </>
   );
 }